+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "gdshader-lsp-client" extension will be documented in
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.0.5]
+
+### Added 
+- GDShader server now compatible from 4.5-4.7+
+- Setting for file logging path
+
 ## [1.0.3]
 
 ### Added